fix(starter): replace long to int

+++ spring-boot-starter/src/main/java/dev/caceresenzo/gitbook/spring/boot/autoconfigure/GitBookProperties.java
@@ -13,6 +13,6 @@ public class GitBookProperties {
 
 	private String apiUrl;
 	private String accessToken;
-	private Long maxPageSize;
+	private Integer maxPageSize;
 
 }
